src/Streamlit/requirements.txt
pandas
seaborn
matplotlib
numpy
category_encoders
sklearn.model_selection
sklearn.linear_model
sklearn.preprocessing